refactor : subscriberCount int로 타입 변경

+++ monew-api/src/main/java/com/monew/monew_api/interest/dto/response/InterestDto.java
@@ -6,7 +6,7 @@ public record InterestDto(
     Long id,
     String name,
     List<String> keywords,
-    Long subscriberCount,
+    int subscriberCount,
     boolean subscribedByMe
 ) {
